debug

+++ database/migrations/2019_08_14_075638_create_table_sfdmx_table.php
@@ -14,7 +14,7 @@ class CreateTableSfdmxTable extends Migration {
 	{
 		Schema::create('table_sfdmx', function(Blueprint $table)
 		{
-			$table->integer('id', true);
+			$table->bigIncrements('id', true);
 			$table->string('sfdh', 20);
 			$table->string('xsbh', 50);
 			$table->string('sfxm', 50);
