+++ database/migrations/2019_08_14_075638_create_table_sfjl_table.php
@@ -14,7 +14,7 @@ class CreateTableSfjlTable extends Migration {
 	{
 		Schema::create('table_sfjl', function(Blueprint $table)
 		{
-			$table->integer('id', true);
+			$table->bigIncrements('id', true);
 			$table->string('sfdh', 20);
 			$table->string('xsbh', 50);
 			$table->decimal('sfje', 10);
